requirements.txt
python-liquid
jsonschema